Adapts the wps classes to make the WpsClient be published as an OSGI service and to prepare for the wps script plugin.
