+++ bundles/wps-client/src/main/java/org/orbisgis/wpsclient/view/ui/dataui/EnumerationUI.java
@@ -25,6 +25,7 @@ import net.opengis.wps._2_0.InputDescriptionType;
 import net.opengis.wps._2_0.OutputDescriptionType;
 import org.orbisgis.sif.common.ContainerItem;
 import org.orbisgis.wpsclient.WpsClient;
+import org.orbisgis.wpsclient.WpsClientImpl;
 import org.orbisgis.wpsclient.view.utils.ToolBoxIcon;
 import org.orbisgis.wpsservice.model.Enumeration;
 import org.slf4j.LoggerFactory;
@@ -60,10 +61,10 @@ public class EnumerationUI implements DataUI{
     private static final String LIST_PROPERTY = "LIST_PROPERTY";
 
     /** WpsClient using the generated UI. */
-    private WpsClient wpsClient;
+    private WpsClientImpl wpsClient;
 
     @Override
-    public void setWpsClient(WpsClient wpsClient){
+    public void setWpsClient(WpsClientImpl wpsClient){
         this.wpsClient = wpsClient;
     }
