+++ bundles/wps-client/src/main/java/org/orbisgis/wpsclient/view/ui/dataui/RawDataUI.java
@@ -26,6 +26,7 @@ import net.opengis.wps._2_0.OutputDescriptionType;
 import org.orbisgis.sif.UIFactory;
 import org.orbisgis.sif.components.OpenPanel;
 import org.orbisgis.wpsclient.WpsClient;
+import org.orbisgis.wpsclient.WpsClientImpl;
 import org.orbisgis.wpsclient.view.utils.ToolBoxIcon;
 import org.orbisgis.wpsservice.model.RawData;
 import org.slf4j.LoggerFactory;
@@ -66,10 +67,10 @@ public class RawDataUI implements DataUI {
     private static final String MULTI_SELECTION_PROPERTY = "MULTI_SELECTION_PROPERTY";
 
     /** WpsClient using the generated UI. */
-    private WpsClient wpsClient;
+    private WpsClientImpl wpsClient;
 
     @Override
-    public void setWpsClient(WpsClient wpsClient){
+    public void setWpsClient(WpsClientImpl wpsClient){
         this.wpsClient = wpsClient;
     }
